playlist service setup
