+++ src/App.js
@@ -1,8 +1,13 @@
+import { useEffect } from 'react';
 import { availableRoutes } from './frontend/routes/routes';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 function App() {
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
+
   return (
     <div className=''>
       {availableRoutes}
