+++ src/frontend/pages/playlist/playlist.css
@@ -1,0 +1,2 @@
+@import '../../utility/styles.css';
+@import '../liked/liked.css';
